test: adding more test and refactoring some names
