+++ src/App.tsx
@@ -7,7 +7,7 @@ import {
 } from "./adapters";
 import { CreditAgreement } from "./models";
 import Dropdown from "./components/DropDown";
-import { Menu } from "./components";
+import { Header } from "./components";
 
 interface AppProps {
   totalWithTax?: number;
@@ -48,7 +48,7 @@ const App: React.FC<AppProps> = ({ totalWithTax = 0 }) => {
 
   return (
     <div className="flex flex-col p-6 border border-gray-400 rounded-l gap-y-2 ">
-      <Menu />
+      <Header />
       <Dropdown options={creditAgreements} />
     </div>
   );
