+++ src/components/Popup.tsx
@@ -24,6 +24,7 @@ const Popup: React.FC<PopupProps> = ({ onClose }) => {
   return (
     <div className="fixed inset-0 flex items-center justify-center z-50">
       <div
+        aria-label="Cerrar mas info"
         className="absolute inset-0 bg-black opacity-50"
         onClick={onClose}
       ></div>
