+++ src/components/__tests__/Header.test.tsx
@@ -2,8 +2,8 @@ import { render, screen } from "@testing-library/react";
 import { describe, it, expect } from "vitest";
 import App from "@/App";
 
-describe("Menu Component", () => {
-  it("renders correctly", async () => {
+describe("Header Component", () => {
+  it("should show the component properly", async () => {
     render(<App />);
 
     const menuElement1 = await screen.findByText("Págalo en");
