+++ src/components/index.ts
@@ -1,3 +1,3 @@
 export { default as DropDown } from "./DropDown";
-export { default as Menu } from "./Menu";
+export { default as Header } from "./Header";
 export { default as Popup } from "./Popup";
